Se agregaron middlewares a las rutas de room, al igual se agregaron comprobaciones de usuairo en las funciones para room

+++ src/controller/room.controller.js
@@ -1,12 +1,15 @@
 'use strict'
 
+const { validateManagerHotel } = require('../helpers/validateManagerHotel');
 const Hotels = require('../models/hotel.model');
 const Room  = require('../models/room.model');
 
 const createRoom = async(req, res) =>{
     try {
-        
+
         const {number,hotel} = req.body;
+
+        if( ! ( validateManagerHotel( req.user._id, hotel ) ) ) return res.status(400).send({ msg: `El usuario logueado no es el manager del hotel.` })
         
         const findHotel = await Hotels.findById(hotel);
         if(!findHotel) return res.status(404).send({message: 'El hotel no se encontro dentro de la base de datos.'})
@@ -43,11 +46,28 @@ const readRooms = async(req, res)=>{
     }
 }
 
+const readRoomsByHotel = async(req, res)=>{
+    try {
+        
+        const {idHotel} = req.body
+        const room = await Room.find( { hotel: idHotel  } )
+
+        if( !room || room.length == 0 ) return res.status(404).send( { msg: `El hotel no cuenta con habitaciones.` } )
+
+        res.status(200).send({message: `Habitaciones del hotel:`, room});
+
+    } catch (error) {
+        console.error(error);
+    }
+}
+
 const updateRoom = async(req, res)=>{
     try {
         
         let {number, hotel} = req.body;
         const {id} = req.params;
+
+        if( ! ( validateManagerHotel( req.user._id, hotel ) ) ) return res.status(400).send({ msg: `El usuario logueado no es el manager del hotel.` })
 
         //Verificar si existe el hotel nuevo
         
@@ -83,6 +103,9 @@ const deleteRoom = async(req, res )=>{
         const {id} = req.params;
 
         const findRoom = await Room.findById(id);
+        
+        if( ! ( validateManagerHotel( req.user._id, findRoom.hotel ) ) ) return res.status(400).send({ msg: `El usuario logueado no es el manager del hotel.` })
+
         if(!findRoom) return res.status(404).send({message: `No se ha encontrado la habitacion.`})
 
         const roomDelete = await Room.findByIdAndDelete({_id: id});
@@ -96,4 +119,4 @@ const deleteRoom = async(req, res )=>{
     }
 }
 
-module.exports = {createRoom,readRooms,updateRoom,deleteRoom};
+module.exports = {createRoom,readRooms, readRoomsByHotel,updateRoom,deleteRoom};
